complete registration page using mern stack

+++ backend/controller/controller.js
@@ -1,12 +1,13 @@
 const asyncHandler = require('express-async-handler')
 
 const Goal = require('../model/goalsModel')
+const User = require("../model/userModel")
 
 // @desc get goals
 // @route Get/api/goals
 // @access private
 const getGoals =asyncHandler (async(req, res)=>{
-    const goals = await Goal.find()
+    const goals = await Goal.find({user:req.user.id})
   
     res.status(200).json(goals)
 })
@@ -21,7 +22,8 @@ const setGoals =asyncHandler( async(req, res)=>{
     }
 
     const goal = await Goal.create({
-        Text: req.body.text
+        Text: req.body.text,
+        User:req.user.id
     })
 
 
@@ -38,6 +40,21 @@ const updateGoals =asyncHandler(async(req, res)=>{
         res.status(400)
         throw new Error('goal not found')
     }
+    const user = await User.findById(red.user.id)
+    // check for user
+    if(!user){
+
+        res.status(401)
+        throw new Error("user not found")
+    }
+    // make sure the logged in user matches goal user
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error("user not authorized")
+
+    }
+
+
     const updateGoal = await Goal.findById(req.params.id,
         req.body, {
             new: true
@@ -54,6 +71,19 @@ const deleteGoals =asyncHandler(async(req, res)=>{
     if(!goal){
         res.status(400)
         throw new Error('goal not found')
+    }
+    const user = await User.findById(red.user.id)
+    // check for user
+    if(!user){
+
+        res.status(401)
+        throw new Error("user not found")
+    }
+    // make sure the logged in user matches goal user
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error("user not authorized")
+
     }
     await goal.deleteOne()
